+++ Nova/Data Management/RecordedMenstrualEventInfo.swift
@@ -11,7 +11,7 @@ import Foundation
 struct RecordedMenstrualEventInfo {
     let sample: MenstrualSample?
     let date: Date
-    let volume: Int
+    let volume: Double
     let selectionState: SelectionState
 }
 
@@ -23,7 +23,7 @@ extension RecordedMenstrualEventInfo: RawRepresentable {
 
     init?(rawValue: RawValue) {
         guard rawValue["name"] as? String == RecordedMenstrualEventInfo.name, let date = rawValue["date"] as? Date,
-            let volume = rawValue["volume"] as? Int,
+            let volume = rawValue["volume"] as? Double,
             let selection = (rawValue["selectionState"] as? SelectionState.RawValue).flatMap(SelectionState.init(rawValue:)) else
         {
             return nil
